app/requirements.txt
fastapi
uvicorn
librosa
numpy
scipy
sentence-transformers
faiss-cpu
python-multipart
pydantic
openai
